Adding otel collector components to scrape metrics
DISCLAMER - This does not work yet

Have added relevant OpenTelemetry collector contrib config files and information to begin scraping metrics from the kafka cluster. Currently only the kafkametrics receiver can work if started alone.

------
ISSUE
------
The main issue that is currently taking place is that the jmx receiver component of the collector seems to try to create a temp ".properties" file in "/tmp" but fails at doing so because with a "permission" denied error, even though an attempt has been made to redirect the temp files to an empty docker volume called "otel-temp-data"